Wrap ws.Server so that versions does not clash
