+++ server.js
@@ -1,0 +1,25 @@
+
+var inherits = require('inherits')
+var WebSocketServer = require('ws').Server
+var stream = require('./stream')
+
+module.exports = Server
+
+function Server(opts) {
+  if (!(this instanceof Server)) {
+    return new Server(opts)
+  }
+
+  WebSocketServer.call(this, opts)
+
+  this.on('newListener', function(event) {
+    if (event === 'stream') {
+      this.on('connection', function(conn) {
+        this.emit('stream', stream(conn))
+      })
+    }
+  })
+}
+
+inherits(Server, WebSocketServer)
+
